requirements.txt
PySide6
Pillow
imageio-ffmpeg
